+++ src/main/java/wooteco/subway/admin/domain/line/Line.java
@@ -1,10 +1,15 @@
-package wooteco.subway.admin.domain;
-
-import org.springframework.data.annotation.Id;
+package wooteco.subway.admin.domain.line;
 
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+
+import org.springframework.data.annotation.Id;
 
 public class Line {
     @Id
