+++ src/main/java/wooteco/subway/admin/domain/line/LineStations.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.line;
 
 import static java.util.stream.Collectors.*;
