+++ src/main/java/wooteco/subway/admin/domain/line/path/EdgeWeightType.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.line.path;
 
 public enum EdgeWeightType {
     DISTANCE("distance",
