+++ src/main/java/wooteco/subway/admin/domain/line/path/SubwayRoute.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.line.path;
 
 import java.util.List;
