+++ src/main/java/wooteco/subway/admin/domain/station/Station.java
@@ -1,8 +1,8 @@
-package wooteco.subway.admin.domain;
-
-import org.springframework.data.annotation.Id;
+package wooteco.subway.admin.domain.station;
 
 import java.time.LocalDateTime;
+
+import org.springframework.data.annotation.Id;
 
 public class Station {
     @Id
