+++ src/main/java/wooteco/subway/admin/dto/StationCreateRequest.java
@@ -1,7 +1,7 @@
 package wooteco.subway.admin.dto;
 
 
-import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.domain.station.Station;
 
 public class StationCreateRequest {
     private String name;
