+++ src/main/java/wooteco/subway/admin/dto/StationResponse.java
@@ -1,10 +1,10 @@
 package wooteco.subway.admin.dto;
 
-import wooteco.subway.admin.domain.Station;
-
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
+
+import wooteco.subway.admin.domain.station.Station;
 
 public class StationResponse {
     private Long id;
