+++ src/main/java/wooteco/subway/admin/repository/StationRepository.java
@@ -1,12 +1,13 @@
 package wooteco.subway.admin.repository;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
-import wooteco.subway.admin.domain.Station;
 
-import java.util.List;
-import java.util.Optional;
+import wooteco.subway.admin.domain.station.Station;
 
 public interface StationRepository extends CrudRepository<Station, Long> {
     @Override
