+++ src/main/resources/static/service/api/index.js
@@ -33,7 +33,7 @@ const api = (() => {
 
   const path = {
     find(params) {
-      return request(`/paths?source=${params.source}&target=${params.target}&type=${params.type}`)
+      return request(`/lines/path/${params.source}/${params.target}`)
     }
   }
