+++ src/test/java/wooteco/subway/admin/domain/line/LineStationTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.line;
 
 import static org.assertj.core.api.Assertions.*;
