+++ src/test/java/wooteco/subway/admin/domain/line/path/EdgeWeightTypeTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.line.path;
 
 import static org.assertj.core.api.Assertions.*;
