+++ src/test/java/wooteco/subway/admin/domain/line/path/SubwayRouteFactoryTest.java
@@ -1,10 +1,12 @@
-package wooteco.subway.admin.domain;
+package wooteco.subway.admin.domain.line.path;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
 import com.google.common.collect.Sets;
+import wooteco.subway.admin.domain.line.LineStation;
+import wooteco.subway.admin.domain.line.LineStations;
 
 class SubwayRouteFactoryTest {
