+++ src/test/java/wooteco/subway/admin/repository/StationRepositoryTest.java
@@ -1,12 +1,13 @@
 package wooteco.subway.admin.repository;
 
+import static org.junit.jupiter.api.Assertions.*;
+
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.data.jdbc.DataJdbcTest;
 import org.springframework.data.relational.core.conversion.DbActionExecutionException;
-import wooteco.subway.admin.domain.Station;
 
-import static org.junit.jupiter.api.Assertions.assertThrows;
+import wooteco.subway.admin.domain.station.Station;
 
 @DataJdbcTest
 public class StationRepositoryTest {
